% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CellMarkersDB}
\alias{CellMarkersDB}
\title{CellMarker database of gene markers}
\format{data frame}
\usage{
data(CellMarkersDB)
}
\description{
database of gene markers downloaded from https://academic.oup.com/nar/article/47/D1/D721/5115823
}
\keyword{cell-markers,}
\keyword{database}
\keyword{datasets,}
