% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{panglaoDB}
\alias{panglaoDB}
\title{PanglaoDB of gene markers}
\format{data frame}
\usage{
data(panglaoDB)
}
\description{
database of gene markers downloaded from https://panglaodb.se/markers.html?cell_type=%27all_cells%27#
(Last updated: 26/06/2019 05:45:59 CET)
}
\keyword{cell-markers,}
\keyword{database}
\keyword{datasets,}
