% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcit_guess_seurat.R, R/wcit_magic.R
\name{wcit_guess_seurat}
\alias{wcit_guess_seurat}
\title{An integration of the wcit package to the Seurat framework}
\usage{
wcit_guess_seurat(seurat_object, number_of_markers = 20)

wcit_guess_seurat(seurat_object, number_of_markers = 20)
}
\arguments{
\item{number_of_markers}{Number of markers extracted from the FindAllMarkers
function that will be used for calculation. Defaults to 20.}

\item{seurat_markers}{A dataframe returned by Seurat's 3.0 FindAllMarkers function}

\item{search_literature}{If true, it will perform a literature search to indicate which,
among the 3 top hits, is most likely to be the one of interest. Defaults to FALSE.}

\item{seurat_markers}{A dataframe returned by Seurat's 3.0 FindAllMarkers function}

\item{number_of_markers}{Number of markers extracted from the FindAllMarkers
function that will be used for calculation. Defaults to 20.}

\item{search_literature}{If true, it will perform a literature search to indicate which,
among the 3 top hits, is most likely to be the one of interest. Defaults to FALSE.}
}
\description{
This function allows you to infer the Seurat clusters identity from a list of putative markers

This function allows you to infer the Seurat clusters identity from a list of putative markers
}
\keyword{single-cell}
