% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcit_magic.R
\name{wcit_magic}
\alias{wcit_magic}
\title{An integration of the wcit package to the Seurat framework}
\usage{
wcit_magic(seurat_object, number_of_markers = 10)
}
\arguments{
\item{seurat_object}{A seurat object already clusterized.}

\item{number_of_markers}{Number of markers extracted from the FindAllMarkers
function that will be used for calculation. Defaults to 10.}
}
\description{
This function allows you to infer the Seurat clusters identity from a list of putative markers
}
\keyword{single-cell}
