% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcit.R
\name{wcit}
\alias{wcit}
\title{A "which cell is that" function.}
\usage{
wcit(markers, panglao = TRUE, CellMarkers = TRUE)
}
\arguments{
\item{markers}{A vector containing the gene markers in Gene Symbol format, for humans}

\item{panglao}{A boolean determining if panglaoDB is going to be used or not. Defaults to TRUE}

\item{CellMarkers}{A boolean determining if CellMarkersDB is going to be used or not. Defaults to TRUE}
}
\value{
A list of putative cell types
}
\description{
This function allows you to infer the cell identity from a list of putative markers
}
\examples{
# B cell markers
  markers <- c("CD79A", "CD79B", "MS4A1")
  marker_hits <- wcit(markers)
}
\keyword{single-cell}
