% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcit.R
\name{wcit}
\alias{wcit}
\title{A "which cell is that" function.}
\usage{
wcit(markers, panglao = T)
}
\arguments{
\item{markers}{A vector containing the gene markers in Gene Symbol format, for humans}

\item{panglao}{A boolean determining if panglaoDB is going to be used or not. Defaults to TRUE}
}
\description{
This function allows you to infer the cell identity from a list of putative markers
}
\examples{
set.seed(3)

data(panglaoDB)
markers <- panglaodb$official.gene.symbol[c(1:10, sample(x = 1:7000,size = 10))]
wcit(markers)
}
\keyword{single-cell}
